git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@10709 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/descent.h
@@ -2212,7 +2212,7 @@ class CSoundData {
 class CTextureData {
 	public:
 		CArray<tBitmapFile>		bitmapFiles [2]; //[MAX_BITMAP_FILES];
-		CArray<sbyte>				bitmapFlags [2]; //[MAX_BITMAP_FILES];
+		CArray<ushort>				bitmapFlags [2]; //[MAX_BITMAP_FILES];
 		CArray<CBitmap>			bitmaps [2]; //[MAX_BITMAP_FILES];
 		CArray<CBitmap>			altBitmaps [2]; //[MAX_BITMAP_FILES];
 		CArray<CBitmap>			addonBitmaps ; //[MAX_ADDON_BITMAP_FILES];
